Completed Prototype
THe search bar now filters out all the other rows but the first row of
titles.
The Background image shading was toned down in the Grayscale css file.
Updated the README to tell the TA/User how to get to our site.
Added in some more text to make the website more descriptive.